+++ tests/test_raster.py
@@ -1,9 +1,12 @@
-import os
 import pathlib
 
-import ptolemy as pt
+import numpy as np
+import pandas as pd
 import pytest
 import xarray as xr
+
+import ptolemy as pt
+
 
 URL = "https://d2ad6b4ur7yvpq.cloudfront.net/naturalearth-3.3.0/ne_110m_admin_0_countries.geojson"
 DATA_PATH = pathlib.Path(__file__).parent / "test_data"
@@ -12,12 +15,6 @@ LIKE["lon"] = LIKE["lon"] - 360
 RASTER_STRATEGIES = ["all_touched", "centroid", "hybrid", "majority", "weighted"]
 
 
-# yes this is silly, but needed until we get a test without lfs
-def test_foo():
-    assert 42 == 42
-
-
-@pytest.mark.lfs
 @pytest.mark.parametrize("as_file", [True, False])
 def test_init(as_file):
     like = LIKE
@@ -31,7 +28,6 @@ def test_init(as_file):
     assert r.coords["lon"].equals(like["lon"])
 
 
-@pytest.mark.lfs
 @pytest.mark.parametrize("flatten,exp_size", [(True, 1), (False, 177)])
 @pytest.mark.parametrize("idxkey", [None, "iso_a3"])
 def test_read_shpf(flatten, exp_size, idxkey):
@@ -48,16 +44,51 @@ def _do_rasterize(strategy):
     return idxr
 
 
-@pytest.mark.lfs
 @pytest.mark.parametrize("strategy", RASTER_STRATEGIES)
 def test_rasterize(strategy):
-    r = pt.Rasterize(like=LIKE)
-    r.read_shpf(URL)
     obs = _do_rasterize(strategy)
     exp = xr.open_dataarray(DATA_PATH / f"{strategy}.nc", mode="r")
     assert obs.equals(exp)
     obs.close()
     exp.close()
+
+
+def test_df_to_raster():
+    r = pt.Rasterize(like=LIKE)
+    r.read_shpf(URL, idxkey="adm0_a3")
+    idxr = r.rasterize(strategy="hybrid", verbose=True)
+    df = pd.DataFrame(
+        {
+            "adm0_a3": ["USA"] * 2 + ["MEX"] * 2,
+            "year": [2015, 2020] * 2,
+            "data": [15, 20, 5, 10],
+        }
+    )
+    idx_map = {v: int(k) for k, v in idxr.attrs.items() if int(k) in np.unique(idxr)}
+    ds = pt.df_to_raster(df, idxr, "adm0_a3", idx_map, coords=["year"])
+    assert ds.sum() == 2855
+
+
+def test_df_to_weighted_raster():
+    r = pt.Rasterize(like=LIKE)
+    r.read_shpf(URL, idxkey="adm0_a3")
+    idxr = r.rasterize(strategy="weighted", verbose=True)
+    df = pd.DataFrame(
+        {
+            "adm0_a3": ["USA"] * 2 + ["MEX"] * 2,
+            "year": [2015, 2020] * 2,
+            "data": [15, 20, 5, 10],
+        }
+    )
+    ds = pt.df_to_weighted_raster(df, idxr, extra_coords=["year"], sum_dim=["adm0_a3"])
+    assert np.isclose(ds.data.sel(year=2015).sum(), 3348.12920833)
+
+
+def test_cell_area():
+    exp = pt.cell_area_from_file(LIKE)
+    assert exp.index[0] > exp.index[1]
+    assert exp.iloc[0] < exp.iloc[1]
+    assert np.isclose(exp.sum(), 1299347051157.3113)
 
 
 if __name__ == "__main__":